GEOMESA-976 Converter Ingest Job and InputFormat
* Supports all converter types for local & mapreduce modes (including binary converters like avro)
* GEOMESA-1033 - Fix Guava version conflict with tools
* GEOMESA-1035 - Define SFT and converter config on command line as files or strings
* GEOMESA-968 - MapReduce counters
* Validated that converted SimpleFeatures have date and time
* Added process(is: InputStream) method to converters (also needed for Nifi)
* Support for multi-line and single-line json/xml documents
* GZ compression for HDFS files
* GZ, XZ, BZ compression support for local files
* Multi-threaded local ingest
* Changed counters to Longs

+++ geomesa-convert/geomesa-convert-text/src/test/scala/org/locationtech/geomesa/convert/text/CompositeTextConverterTest.scala
@@ -80,8 +80,8 @@ class CompositeTextConverterTest extends Specification {
     }
 
     "and get correct line numbers" >> {
-      res(0).getAttribute("lineNr").asInstanceOf[Int] must be equalTo 1
-      res(1).getAttribute("lineNr").asInstanceOf[Int] must be equalTo 4
+      res(0).getAttribute("lineNr").asInstanceOf[Long] must be equalTo 1
+      res(1).getAttribute("lineNr").asInstanceOf[Long] must be equalTo 4
     }
 
     "testing string2 function defaults" >> {
